feat(deploys): store and display full logs (#224)
Closes #221 

## Changes

- Created a dedicated table to store raw logs
We are not using the main table to limit impact on query time when
listing jobs without needing to read logs.

- Use a double logger to display the log in the console, and store them
in a file
It was the easiest to have the two output. Later, we will store the file
somewhere else (e.g: GCS) and drop the table entirely.
The problem right now is that `multistream` in Pino does not support
formatter, so the output in prod will not be GCP compliant. The ideal
long-term solution is to run each job in a container and output all logs
to a file.

- Sync
Clean up leftover from when this was a CLI
Add tests to cover the basics.

<img width="975" alt="Screenshot 2023-09-15 at 15 38 31"
src="https://github.com/specfy/specfy/assets/1637651/10275946-7671-462c-81c1-1b945fca84ed">