feat(flow): automatic edge placement (#235)
Edge now automatically change start and end position depending on the
Node relative position.
It's handy to reorganise a Flow, especially at initial import and
removes the need to manually update anything.

The only drawback, if it's not exactly what we want it will always
override the manual changes.
We could enable/disable this behavior at the project or user level but
both options are imperfect. An other solution could be to lock edge and
only move the ones that are not locked.