feat: Sources are now outside Projects table (#246)
## Need
- be able to manage more than one data source
- track sourceId for futur features instead of "type + name" 

## Updates
- Create table `Sources`
One row per source, per project. Currently only one will be supported to
be retrocompat.
Now the concerns are well separated.

- Create table `SourceEntries`
N row per component. This will contain all the source + path where this
component was found. This will enable transparency and merging many
sources into one component.

- Double write until deprecation
To make sure the migration can be reverted we keep the old columns and
we keep updating the old columns

- Old data is copied inside the migration
As always with this kind of migration, it could miss some data but that
would be very unlucky.

- Link project to GitHub now handled by /sources
Sources is now the main model to link project. Currently only supporting
one kind.