feat: process events asynchronously  (#379)
Some critical paths, like registering, were slowed down by side tasks
(e.g: emailing, reaching third parties, etc.).

A simple solution, but not perfect is to move those tasks outside the
critical path. Not perfect because they still are in the same NodeJs
event loop; so they might still block the process but at least the API
can return before being blocked.

The event emitter is a simple in-memory dispatcher, so in case of crash
, or early exit, we can loose events but it should be an acceptable edge
case for the moment.